% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaWeightedTotal.R
\name{areaWeightedTotal}
\alias{areaWeightedTotal}
\title{Calculate area-weighted total}
\usage{
areaWeightedTotal(
  x,
  area_unit = "m^2",
  lon_centres = seq(-179.75, 179.75, by = 0.5),
  lat_centres = seq(-89.75, 89.75, by = 0.5)
)
}
\arguments{
\item{x}{A spatial Field for which the area-weighted summed flux}

\item{area_unit}{A character string specifying the unit the Field is in, default is "m^2" but can also be "km^2" or "ha"
(see DGVMTools::addArea)}
}
\value{
A DGVMTools::Field object the area-weighted summed data.
}
\description{
Calculates area-weighted sum of a Field.  Useful for deriving global/regional GPP or other fluxes, C stocks etc.
Note the Field variable must be on a per-unit area basis, with the area unit supplied as an argument (default is "m^2").
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
