% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-ICOS.R
\name{availableQuantities_ICOS}
\alias{availableQuantities_ICOS}
\title{List ICOS quantities available}
\usage{
availableQuantities_ICOS()
}
\value{
A list of all ICOS quantities
}
\description{
Simply lists all carbon flux variables available in the ICOS data
}
\keyword{internal}
