% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availiableICOSStations.R
\name{availiableICOSStations}
\alias{availiableICOSStations}
\title{List all availiable ICOS stations and get Lat/Lon corrections.}
\usage{
availiableICOSStations(Stations = "All", station.selection = "Hyltemossa")
}
\arguments{
\item{Stations}{Default "ALL" renders datatable with all availiable stations.}

\item{station.selection}{Make a selection by station name to get correct Lat/Lon. Takes singe chr vector or list of vectors.}
}
\value{
datatable with station information.
}
\description{
List all availiable ICOS stations and get Lat/Lon corrections.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
