% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSummaryTable.R
\name{buildSummaryTable}
\alias{buildSummaryTable}
\title{Build summary table from benchmark. Provides the spatial mean value of your data sources.}
\usage{
buildSummaryTable(benchmark = this_benchmark, all_sim_full, summary_col_names)
}
\arguments{
\item{benchmark}{Holds current benchmark instructions}

\item{all_sim_full}{list of all simulation fields}

\item{summary_col_names}{Column header names}
}
\value{
datatable which can be knitted into tabular output
}
\description{
Build summary table from benchmark. Provides the spatial mean value of your data sources.
}
\examples{
buildSummaryTable(benchmark = this_benchmark, all_sim_full, summary_col_names)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
