% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUnits.R
\name{convertUnits}
\alias{convertUnits}
\title{Title}
\usage{
convertUnits(df, from, to, benchmark = this_benchmark)
}
\arguments{
\item{df}{The fields you need to convert}

\item{from}{Current unit}

\item{to}{The unit you want to convert to}

\item{benchmark}{Holds information on which layer to convert}
}
\value{
Field with converted units
}
\description{
Title
}
\examples{
convertUnits(Data.year.mean, "indiv/ha", "indiv/km^2")
}
