% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFLUXNETDataset.R
\name{correctFLUXNETDataset}
\alias{correctFLUXNETDataset}
\title{Corrects Coordinates and station names for FLUXNET}
\usage{
correctFLUXNETDataset(Dataset)
}
\arguments{
\item{Dataset}{FLUXNET Dataset}
}
\value{
Corrected Dataset
}
\description{
Corrects Coordinates and station names for FLUXNET
}
\examples{
correctFLUXNETDataset(Dataset = Data)
}
