% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctICOSDatasets.R
\name{correctICOSDataset}
\alias{correctICOSDataset}
\title{Title}
\usage{
correctICOSDataset(all_Fields_list, site.type = NULL)
}
\arguments{
\item{all_Fields_list}{dataset in all_Fields_list}

\item{site.type}{Choose between "ecosystem" stations or "atmospheric" stations}
}
\value{
Corrected Lat Lon for ICOS dataset
}
\description{
Title
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
