% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpoolTable.R
\name{CpoolTable}
\alias{CpoolTable}
\title{CpoolTable}
\usage{
CpoolTable(
  param = NULL,
  benchmark = this_benchmark,
  all_sim_full = all_sim_full
)
}
\arguments{
\item{param}{Not used}

\item{benchmark}{the current benchmark object}

\item{all_sim_full}{List of simulations}
}
\value{
Table with cpool Litter, Soil & AGB per country
}
\description{
CpoolTable
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
