% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBenchmark.R
\name{createBenchmark}
\alias{createBenchmark}
\title{Title}
\usage{
createBenchmark(
  benchmark_name,
  input,
  simulation,
  Data.year.mean,
  spatial.extent = NULL
)
}
\arguments{
\item{benchmark_name}{The name of the current benchmark,
should follow the benchmark object name from ins file.}

\item{input}{YAML ins file}

\item{simulation}{Specifies the simulation to work with.
Value should be integer specifying the correct simulation name from the list in ins.}

\item{Data.year.mean}{The dataset defined previously in benchmark panel,
can be empty and should be left as is.}

\item{spatial.extent}{The spatial extent set through yaml ins,
default is NULL which use the full extent of your fields.}
}
\value{
the current benchmark class on which further analysis will be based.
}
\description{
Title
}
\examples{
this_benchmark <- createBenchmark(benchmark_name, input, simulation = 2, Data.year.mean, spatial.extent)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
