% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_aDGVM2_Sources.R
\name{define_aDGVM2_Sources}
\alias{define_aDGVM2_Sources}
\title{Define aDGVM2 sources. Create a aDGVM2 source for your dataset and simulations.}
\usage{
define_aDGVM2_Sources(input)
}
\arguments{
\item{input}{}
}
\value{
List of lists for the dataset source and the simulations sources
}
\description{
Define aDGVM2 sources. Create a aDGVM2 source for your dataset and simulations.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
