% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_aDGVM_Sources.R
\name{define_aDGVM_Sources}
\alias{define_aDGVM_Sources}
\title{Define aDGVM sources. Create a aDGVM source for your dataset and simulations.}
\usage{
define_aDGVM_Sources(input)
}
\arguments{
\item{input}{}
}
\value{
List of lists for the dataset source and the simulations sources
}
\description{
Define aDGVM sources. Create a aDGVM source for your dataset and simulations.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
