% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_FLUXNET_DatasetSource.R
\name{define_FLUXNET_DatasetSource}
\alias{define_FLUXNET_DatasetSource}
\title{Define FLUXNET sources. Create a FLUXNET source for your dataset.}
\usage{
define_FLUXNET_DatasetSource(input)
}
\arguments{
\item{input}{YML input to get your directory information}
}
\value{
List of lists
}
\description{
Define FLUXNET sources. Create a FLUXNET source for your dataset.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
