% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_GUESS_Sources.R
\name{define_GUESS_Sources}
\alias{define_GUESS_Sources}
\title{Define GUESS sources. Create a GUESS source for your dataset and simulations.}
\usage{
define_GUESS_Sources(input)
}
\arguments{
\item{input}{}
}
\value{
List of lists for the dataset source and the simulations sources
}
\description{
Define GUESS sources. Create a GUESS source for your dataset and simulations.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
