% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_ICOS_DatasetSource.R
\name{define_ICOS_DatasetSource}
\alias{define_ICOS_DatasetSource}
\title{Define ICOS sources. Create a ICOS source for your dataset.}
\usage{
define_ICOS_DatasetSource(input)
}
\arguments{
\item{input}{YML input to get your directory information}
}
\value{
List of lists
}
\description{
Define ICOS sources. Create a ICOS source for your dataset.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
