% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_NetCDF_Sources.R
\name{define_NetCDF_Sources}
\alias{define_NetCDF_Sources}
\title{Define NetCDF sources. Create a NetCDF source for your dataset and simulations.}
\usage{
define_NetCDF_Sources(input)
}
\arguments{
\item{input}{YML input to get your directory information}
}
\value{
List of lists for the dataset source and the simulations sources
}
\description{
Define NetCDF sources. Create a NetCDF source for your dataset and simulations.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
