% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_SITE_Sources.R
\name{define_SITE_Sources}
\alias{define_SITE_Sources}
\title{Define SITE sources. Create a SITE source for your dataset and simulations.}
\usage{
define_SITE_Sources(input)
}
\arguments{
\item{input}{}
}
\value{
List of lists for the dataset source and the simulations sources
}
\description{
Define SITE sources. Create a SITE source for your dataset and simulations.
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
