% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAllQuantities.R
\name{defineAllQuantities}
\alias{defineAllQuantities}
\title{Define all availiable quanteties and read correct format}
\usage{
defineAllQuantities(input)
}
\arguments{
\item{input}{Input yaml instruction file}
}
\value{
Updated quantity list for specific format
}
\description{
Define all availiable quanteties and read correct format
}
\examples{
defineAllQuantities(input = input_data)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
