% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestFraction.R
\name{ForestFraction}
\alias{ForestFraction}
\title{ForestFraction. Build table of forest fraction per country. Can be used for one benchmark but can also be continually updated by using rbind() when calling function.}
\usage{
ForestFraction(benchmark = this_benchmark, all_Fields_list = all_Fields_list)
}
\arguments{
\item{benchmark}{The current benchmark object}

\item{all_Fields_list}{The list holding all your fields}
}
\description{
ForestFraction. Build table of forest fraction per country. Can be used for one benchmark but can also be continually updated by using rbind() when calling function.
}
\examples{
ForestFraction(benchmark = this_benchmark, all_Fields_list = all_Fields_list)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
