% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSpatialComparison.R,
%   R/fullTemporalComparison.R
\name{fullSpatialComparison}
\alias{fullSpatialComparison}
\alias{fullTemporalComparison}
\title{Full Spatial Comparison}
\usage{
fullSpatialComparison(
  benchmark,
  all_maps,
  all_trends = NULL,
  all_seasonal = NULL,
  reference_simulation = NULL
)

fullTemporalComparison(
  benchmark,
  all_ts,
  all_trends = NULL,
  all_seasonal = NULL,
  reference_simulation = NULL
)
}
\arguments{
\item{benchmark}{An object of class Benchmark which described the benchmark to be performed.  Needed only for its "guess_layers" and "datasets" slots.}

\item{all_maps}{A list of all spatial values Fields to be compared (both data and simulations, which of these are data are identified from the benchmark argument)}

\item{all_trends}{A list of all trend Fields to be compared (structure as above))}

\item{all_seasonal}{A list of all seasonal Fields to be compared (structure as above))}

\item{reference_simulation}{Characters strings defining the reference model run (for "new minus reference" style comparisons)}

\item{all_ts}{A list of all spatial values Fields to be compared (both data and simulations, which of these are data are identified from the benchmark argument)}
}
\value{
A list'o'lists

A list'o'lists
}
\description{
SOme blah

SOme blah
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
