% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllFields.R
\name{getAllFields}
\alias{getAllFields}
\title{GetAllFields, processing the actual fields from your sources based on the benchmark instructions}
\usage{
getAllFields(benchmark = this_benchmark, all_simulation_Sources_list)
}
\arguments{
\item{benchmark}{Holding current benchmark instructions for getfield call}

\item{all_simulation_Sources_list}{list of all simulation sources}
}
\value{
list of lists containing all simulation fields and all fields including data
}
\description{
GetAllFields, processing the actual fields from your sources based on the benchmark instructions
}
\examples{
getAllFields(benchmark = this_benchmark, all_simulation_Sources_list = all_simulation_Sources_list)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
