% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-FLUXNET.R
\name{getField_FLUXNET}
\alias{getField_FLUXNET}
\title{Get a Field for FLUXNET}
\usage{
getField_FLUXNET(
  source,
  quant,
  layers = NULL,
  target.STAInfo,
  file.name,
  verbose,
  UT.threshold = "VUT",
  partition.method = "REF",
  day.night.method = "NT",
  first.year,
  last.year,
  ...
)
}
\arguments{
\item{source}{A  \code{\linkS4class{Source}} containing the meta-data about the FLUXNET observations}

\item{quant}{A Quantity object to define what quantity from the FLUXNET observations to extract}

\item{layers}{Ignored for FLUXNET}

\item{target.STAInfo}{The spatial-temporal target domain}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is just taken to be
"\href{mailto:quant@id}{quant@id}.out" (also "\href{mailto:quant@id}{quant@id}.out.gz")}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{UT.threshold}{USTAR threshold, variable (VUT) or constant (CUT). VUT is default}

\item{partition.method}{Partition method, e.g. REF, USTAR50, MEAN. REF is default}

\item{day.night.method}{Nighttime (NT) or daytime (DT) partition method for GPP and Reco. Default is NT}

\item{first.year}{Optional, will exclude data before this year}

\item{last.year}{Optional, will exclude data beyond this year}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from FLUXNET observations.
}
\author{
Margot Knapen \email{margot.knapen@nateko.lu.se}
}
\keyword{internal}
