% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-ICOS.R
\name{getField_ICOS}
\alias{getField_ICOS}
\title{Get a Field for ICOS}
\usage{
getField_ICOS(
  source,
  quant,
  layers = NULL,
  path.ICOS,
  UT.threshold = "VUT",
  partition.method = "REF",
  first.year = "",
  last.year = "",
  verbose = F
)
}
\arguments{
\item{source}{A  \code{\linkS4class{Source}} containing the meta-data about the ICOS observations}

\item{quant}{A Quantity object to define what quantity from the ICOS observations to extract}

\item{layers}{Ignored for ICOS}

\item{path.ICOS}{Path to where the ICOS files are stored}

\item{UT.threshold}{USTAR threshold, variable (VUT) or constant (CUT). VUT is default}

\item{partition.method}{Partition method, e.g. REF, USTAR50, MEAN. REF is default}

\item{first.year}{Optional, will exclude data before this year}

\item{last.year}{Optional, will exclude data beyond this year}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from ICOS observations.
}
\author{
Margot Knapen \email{margot.knapen@nateko.lu.se}
}
\keyword{internal}
