% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-SITE.R
\name{getField_SITE}
\alias{getField_SITE}
\title{Get a Field for ICOS}
\usage{
getField_SITE(
  source,
  quant,
  layers = NULL,
  target.STAInfo,
  file.name,
  first.year,
  last.year,
  verbose,
  ...
)
}
\arguments{
\item{source}{A  \code{\linkS4class{Source}} containing the meta-data about the ICOS observations}

\item{quant}{A Quantity object to define what quantity from the ICOS observations to extract}

\item{layers}{Ignored for SITE}

\item{target.STAInfo}{The spatial-temporal target domain}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is just taken to be
"\href{mailto:quant@id}{quant@id}.out" (also "\href{mailto:quant@id}{quant@id}.out.gz")}

\item{first.year}{Optional, will exclude data before this year}

\item{last.year}{Optional, will exclude data beyond this year}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from ICOS observations.
}
\author{
Karl Piltz \email{karl.piltz@nateko.lu.se}
}
\keyword{internal}
