% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestTable.R
\name{HarvestTable}
\alias{HarvestTable}
\title{Harvest table}
\usage{
HarvestTable(param, benchmark = this_benchmark, all_sim_full = all_sim_full)
}
\arguments{
\item{param}{What metrics to produce, either "Total Harvest" for harvest volume metrics or "NAI" for net annual increment metrics.}

\item{benchmark}{The current benchmark.}

\item{all_sim_full}{List holding simulations.}
}
\description{
Harvest table
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se), Susanne Suvanto (susanne.suvanto@luke.fi)
}
