% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetricTable.R
\name{makeMetricTable}
\alias{makeMetricTable}
\title{Make metric table lines}
\usage{
makeMetricTable(
  benchmark,
  all_comparisons_list,
  all_simulations_Sources_list,
  signif = 3
)
}
\arguments{
\item{benchmark}{An object of class Benchmark}

\item{col_names}{A vector of character strings describing the column names}
}
\value{
A list which forms the lines of a benchmarking summary table
}
\description{
SOme blah
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
