% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummaryLine.R
\name{makeSummaryLine}
\alias{makeSummaryLine}
\title{Make a summary table line}
\usage{
makeSummaryLine(benchmark, col_names)
}
\arguments{
\item{benchmark}{An object of class Benchmark}

\item{col_names}{A vector of character strings describing the column names}
}
\value{
A list which forms the lines of a benchmarking summary table
}
\description{
SOme blah
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
