% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTemporalComparisons.R
\name{plotAllTemporalComparisons}
\alias{plotAllTemporalComparisons}
\title{Make plots from temporal comparison objects from the difference between simulation and data.}
\usage{
plotAllTemporalComparisons(
  Benchmark = this_benchmark,
  all_comparisons = all_comparisons
)
}
\arguments{
\item{Benchmark}{Holding the current benchmark information.}

\item{all_comparisons}{List holding all comparison objects.}
}
\value{
Temporal difference plots
}
\description{
Make plots from temporal comparison objects from the difference between simulation and data.
}
\examples{
plotAllTemporalComparisons(Benchmark = this_benchmark, all_comparisons = all_comparisons)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
