% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFLUXNETScatter.R
\name{plotFLUXNETScatter}
\alias{plotFLUXNETScatter}
\title{Make FLUXNET scatter plots. Creates scatter plots and use fill to display per Land use or Climate etc.}
\usage{
plotFLUXNETScatter(
  Benchmark = this_benchmark,
  all_comparisons = all_comparisons,
  fill = NULL
)
}
\arguments{
\item{Benchmark}{Holds current becnhmark information.}

\item{all_comparisons}{List holding the comparison objects}

\item{fill}{fill is used to classify points by different characteristics. Use "climate" for koppen climate zones and "landuse" for land cover information, "country" or "station".}
}
\value{
Scatter plots of FLUXNET comparisons.
}
\description{
Make FLUXNET scatter plots. Creates scatter plots and use fill to display per Land use or Climate etc.
}
\examples{
plotFLUXNETScatter(Benchmark = this_benchmark, all_comparisons = all_comparisons, fill = "climate")
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
