% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFullScatter.R
\name{plotFullScatter}
\alias{plotFullScatter}
\title{Plot scatter/density plot of all simulation and dataset}
\usage{
plotFullScatter(Benchmark, all_sim_full)
}
\arguments{
\item{Benchmark}{Holds current becnhmark settings}

\item{all_sim_full_AGPP}{Holds list of all simulations}
}
\value{
Returns refined density plot of simulations vs dataset
}
\description{
Plot scatter/density plot of all simulation and dataset
}
\examples{
plotFullSpatial(Benchmark = this_benchmark, all_sim_full = all_sim_full_AGPP)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
