% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotICOSFields.R
\name{plotICOSFields}
\alias{plotICOSFields}
\title{Plot ICOS dataset and runs either per run or by comparison Model/Data.}
\usage{
plotICOSFields(
  Benchmark = this_benchmark,
  all_Fields_list = all_Fields_list,
  all_comparisons = all_comparisons,
  plot.option
)
}
\arguments{
\item{Benchmark}{Object holding current benchmark information}

\item{all_Fields_list}{The list holding all exclusive fields}

\item{all_comparisons}{The list holding comparison objects for joined plots}

\item{plot.option}{Plotting options. Either "per_source" to make a plot for each source exclusively, or "joined" to make plots of data together with simulations.}
}
\value{
plots of ICOS time.series
}
\description{
Plot ICOS dataset and runs either per run or by comparison Model/Data.
}
\examples{
plotICOSFields(Benchmark = this_benchmark, all_comparisons = all_comparisons, plot.option = "joined") plotICOSFields(Benchmark = this_benchmark, all_Fields_list = all_Fields_list, plot.option = "per_source")
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
