% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotICOSScatter.R
\name{plotICOSScatter}
\alias{plotICOSScatter}
\title{Make ICOS scatter plots. Creates scatter plots and use fill to display per Land use or Climate etc.}
\usage{
plotICOSScatter(
  Benchmark = this_benchmark,
  all_comparisons = all_comparisons,
  fill = NULL
)
}
\arguments{
\item{Benchmark}{Holds current becnhmark information.}

\item{all_comparisons}{List holding the comparison objects}

\item{fill}{fill is used to classify points by different characteristics. Use "climate" for koppen climate zones and "landuse" for land cover information.}
}
\value{
Scatter plots of ICOS comparisons.
}
\description{
Make ICOS scatter plots. Creates scatter plots and use fill to display per Land use or Climate etc.
}
\examples{
plotICOSScatter(Benchmark = this_benchmark, all_comparisons = all_comparisons, fill = "climate")
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
