% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialField.R
\name{plotSpatialField}
\alias{plotSpatialField}
\title{Plot Field or list of fields}
\usage{
plotSpatialField(Benchmark, all_Fields_list = all_Fields_list)
}
\arguments{
\item{Benchmark}{Holds the current benchmark settings}

\item{all_Fields_list}{list containing all your fields both data and simulations}
}
\description{
Plot Field or list of fields
}
