% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStormMap.R
\name{plotStormMap}
\alias{plotStormMap}
\title{plotStormMap}
\usage{
plotStormMap(benchmark = this_benchmark, all_sim_full, LC = NULL)
}
\arguments{
\item{benchmark}{The current benchmark.}

\item{all_sim_full}{List holding simulations.}

\item{LC}{Landcover file read as txt}
}
\description{
plotStormMap
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se), Fredrik Lagergren (fredrik.lagergren@nateko.lu.se)
}
