% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalFields.R
\name{plotTemporalFields}
\alias{plotTemporalFields}
\title{Plotting timeseries with uncertainty metrics.}
\usage{
plotTemporalFields(
  Benchmark = this_benchmark,
  all_Fields_list = all_Fields_list,
  all_comparisons = all_comparisons,
  plot.option,
  uncertainty = "none"
)
}
\arguments{
\item{Benchmark}{The current benchmark information}

\item{all_Fields_list}{Used to plot per source}

\item{all_comparisons}{Used to plut combined timeseries}

\item{plot.option}{Option for pre layer or combined plots use "per_source" or "joined".}

\item{uncertainty}{Default = "none" . Set "SD" to use standard deviation, Looks for ancilliary files.}
}
\value{
plots
}
\description{
Plotting timeseries with uncertainty metrics.
}
\examples{
plotTemporalFields(Benchmark = this_benchmark, all_Fields_list = all_Fields_list, all_comparisons = all_comparisons, plot.option , uncertainty = "SD")
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
