% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalScatter.R
\name{plotTemporalScatter}
\alias{plotTemporalScatter}
\title{Plot temporal scatterplots. Plotting the mean per site with possebility of uncertainty metrics.}
\usage{
plotTemporalScatter(
  Benchmark = this_benchmark,
  all_comparisons = all_comparisons,
  uncertainty = "none"
)
}
\arguments{
\item{Benchmark}{The current benchmark}

\item{all_comparisons}{the list holding comparisons}

\item{uncertainty}{What method of uncertainty should be used "SD" or "Range" default = "none"}
}
\value{
Scatterplot of mean site measurments
}
\description{
Plot temporal scatterplots. Plotting the mean per site with possebility of uncertainty metrics.
}
\examples{
plotTemporalScatter(Benchmark = this_benchmark, all_comparisons = all_comparisons, uncertainty = "SD")
}
\author{
Karl piltz (karl.piltz@nateko.lu.se)
}
