% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input.R, R/read_site_metadata.R
\name{read_site_metadata}
\alias{read_site_metadata}
\title{Read site metadata}
\usage{
read_site_metadata(file_name)

read_site_metadata(file_name)
}
\arguments{
\item{file_name}{Character, an ICOS text file (no processing required)}
}
\value{
A list of S4 objects of class Site.

A list of S4 objects of class Site.
}
\description{
blah

blah
}
\author{
Margot Knapen \email{Margot.Knapen@nateko.lu.se}

details

Margot Knapen \email{Margot.Knapen@nateko.lu.se}

details
}
