% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regrowth.R
\name{Regrowth}
\alias{Regrowth}
\title{Regrowth benchmark. Produces a plot of regrowth using binned growth after disturbance by biome.
Function reads internal gridlist with lon, lat, biome.
Also an internal regrowth dataset containing Lat, Lon, Year, Biome, bin.}
\usage{
Regrowth(
  benchmark = this_benchmark,
  all_sim_full,
  opt.gridlist = NULL,
  opt.dataset = NULL
)
}
\arguments{
\item{benchmark}{Current benchmark used to select column to base calculations on.}

\item{all_sim_full}{List containing your simulations}

\item{opt.gridlist}{Full "path" to gridlist if you provide your own for special cases. Default = NULL.}

\item{opt.dataset}{Full "path" to dataset if you provide your own for special cases. Default = NULL.}
}
\description{
Regrowth benchmark. Produces a plot of regrowth using binned growth after disturbance by biome.
Function reads internal gridlist with lon, lat, biome.
Also an internal regrowth dataset containing Lat, Lon, Year, Biome, bin.
}
\examples{
Regrowth(this_benchmark, all_sim_full)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se) Anne-marie Eckes-Shepard (annemarie.eckes-shephard@nateko.lu.se)
}
