% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGridCellExtent.R
\name{setGridCellExtent}
\alias{setGridCellExtent}
\title{Subset grid cells by cropping by extent}
\usage{
setGridCellExtent(input)
}
\arguments{
\item{input}{Holds information from the instruction yaml file}
}
\value{
List of all cropped fields
}
\description{
Subset grid cells by cropping by extent
}
\examples{
setGridCellExtent(input = input)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
