% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Benchmark-class}
\alias{Benchmark-class}
\title{Benchmark class}
\description{
This class encapsulates all the specific details benchmark comparison.
}
\details{
This class defines metadata about a benchmark and the DGVMTools::Field to which the model output should be compared.  \emph{But} it should be kept in
mind that benchmarks and datasets can vary a lot in terms of units, processing needed, how to deal with different time period etc, so the actual information
stored in the class needs to be used carefully and should not be too automated.
}
\section{Slots}{

\describe{
\item{\code{id}}{Simple character string to gave an uniquely identify this Benchmark}

\item{\code{name}}{More complex character string to describe this Benchmark}

\item{\code{description}}{An even more complex string to describe the Benchmark, probably redundant}

\item{\code{simulation}}{A character string to describe which type of simulation is needed to compare to this benchmark.  Can be a vector, in which they
will be tried in order}

\item{\code{guess_var}}{Character describing the modelled variable(s) needed for this benchmark}

\item{\code{guess_layers}}{Character describing the modelled layers(s) needed for this benchmark}

\item{\code{unit}}{Character string for the unit used in individual gridcells in the spatial plots, will be evaluated as an expression.}

\item{\code{agg.unit}}{Character string for the aggregated units used for global summaries, will be evaluated as an expression.}

\item{\code{datasets}}{A list of DGVMTools::Field objects that are used for comparison in this benchmark}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
