% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLinearTrend.R
\name{calcLinearTrend}
\alias{calcLinearTrend}
\title{Calculate linear trend}
\usage{
calcLinearTrend(field)
}
\arguments{
\item{field}{The Field for which the trend is to be calculated.  Should be spatial, and have a least a "Year" time dimension.}
}
\value{
A DGVMTools::Field object containing the trend and the associated p-value.
}
\description{
Calculate a linear trend from a single variable spatial Field.  Field must have a "Year" dimension, and can optionally be monthly or daily.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
