% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAllSources.R
\name{defineAllSources}
\alias{defineAllSources}
\title{Define all sources}
\usage{
defineAllSources(input, format)
}
\arguments{
\item{input}{yaml instruction file}

\item{format}{Holding format}
}
\value{
List of lists for the dataset source and the simulations sources
}
\description{
Define all sources
}
\examples{
defineAllSources(input = input_data, format = format)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
