% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireTable.R
\name{FireTable}
\alias{FireTable}
\title{FireTable}
\usage{
FireTable(param = NULL, benchmark = this_benchmark, all_sim_full)
}
\arguments{
\item{param}{What metrics to produce, either "BurntFr"  or "Fire".}

\item{benchmark}{The current benchmark.}

\item{all_sim_full}{List holding simulations.}
}
\description{
FireTable
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se), Susanne Suvanto (susanne.suvanto@luke.fi)
}
