% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSpatialComparison.R
\name{fullSpatialComparison}
\alias{fullSpatialComparison}
\title{Full Spatial Comparison}
\usage{
fullSpatialComparison(
  benchmark,
  all_maps,
  all_trends,
  all_seasonal,
  new_model = NULL,
  old_model = NULL
)
}
\arguments{
\item{benchmark}{An object of class Benchmark which described the benchmark to be performed.  Needed only for its "guess_var" and "datasets" slots.}

\item{all_maps}{A list of all spatial values Fields to be compared (both data and simulations, which of these are data are identified from the benchmark argument)}

\item{all_trends}{A list of all trend Fields to be compared (structure as above))}

\item{all_seasonal}{A list of all seasonal Fields to be compared (structure as above))}

\item{new_model, old_model}{Characters strings defing the new and old model runs from a "new minus old" comparisons)}
}
\value{
A list'o'lists
}
\description{
SOme blah
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
