% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllSpatialComparisons.R
\name{plotAllSpatialComparisons}
\alias{plotAllSpatialComparisons}
\title{Plots all comparison objects from fullSpatialComparison as refined difference plots}
\usage{
plotAllSpatialComparisons(Benchmark, all_comparisons, type = "difference")
}
\arguments{
\item{Benchmark}{Benchmark holds the settings for the benchmark currently being performed used to set range parameters to the plotting.}

\item{all_comparisons}{List holding all comparison objects}

\item{type}{A character specifying what type of plot to make. Can be "difference" (default, for a difference plot), "percentage.difference", "values" (actual values, side-by-side).}
}
\value{
Returns difference for every possible comparison in your comparison list.
}
\description{
Plots all comparison objects from fullSpatialComparison as refined difference plots
}
\examples{
plotAllSpatialComparisons(Benchmark = this_benchmark, all_comparisons = all_comparisons)
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se)
}
