% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStormScatter.R
\name{PlotStormScatter}
\alias{PlotStormScatter}
\title{Storm disturbance scatter}
\usage{
PlotStormScatter(benchmark = this_benchmark, all_sim_full)
}
\arguments{
\item{benchmark}{The current benchmark.}

\item{all_sim_full}{List holding simulations.}
}
\value{
Scatterplot of modelled and observed damage
}
\description{
Storm disturbance scatter
}
\author{
Karl Piltz (karl.piltz@nateko.lu.se), Fredrik Lagergren (fredrik.lagergren@nateko.lu.se)
}
